Added some styles to the components
