+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\index.js":"1","C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\App.js":"2","C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\reportWebVitals.js":"3","C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\components\\InitialPage\\InitialPage.js":"4","C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\components\\Playground\\Playground.js":"5","C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\components\\Drums\\Drums.js":"6","C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\components\\Mixer\\Mixer.js":"7","C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\stores\\audioManager.js":"8","C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\sounds.js":"9"},{"size":500,"mtime":499162500000,"results":"10","hashOfConfig":"11"},{"size":464,"mtime":1611092442037,"results":"12","hashOfConfig":"11"},{"size":362,"mtime":499162500000,"results":"13","hashOfConfig":"11"},{"size":506,"mtime":1611094771384,"results":"14","hashOfConfig":"11"},{"size":852,"mtime":1611213390390,"results":"15","hashOfConfig":"11"},{"size":141,"mtime":1611092593460,"results":"16","hashOfConfig":"11"},{"size":1015,"mtime":1611212614502,"results":"17","hashOfConfig":"11"},{"size":2544,"mtime":1611175085645,"results":"18","hashOfConfig":"11"},{"size":796,"mtime":1611135583425,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},"13beoam",{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\index.js",[],["39","40"],"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\App.js",[],"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\reportWebVitals.js",[],"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\components\\InitialPage\\InitialPage.js",[],"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\components\\Playground\\Playground.js",[],"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\components\\Drums\\Drums.js",[],"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\components\\Mixer\\Mixer.js",[],"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\stores\\audioManager.js",["41"],"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\sounds.js",[],{"ruleId":"42","replacedBy":"43"},{"ruleId":"44","replacedBy":"45"},{"ruleId":"46","severity":1,"message":"47","line":43,"column":33,"nodeType":"48","messageId":"49","endLine":43,"endColumn":35},"no-native-reassign",["50"],"no-negated-in-lhs",["51"],"array-callback-return","Array.prototype.map() expects a return value from arrow function.","ArrowFunctionExpression","expectedInside","no-global-assign","no-unsafe-negation"]
+[{"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\index.js":"1","C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\App.js":"2","C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\reportWebVitals.js":"3","C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\components\\InitialPage\\InitialPage.js":"4","C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\components\\Playground\\Playground.js":"5","C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\components\\Drums\\Drums.js":"6","C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\components\\Mixer\\Mixer.js":"7","C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\stores\\audioManager.js":"8","C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\sounds.js":"9"},{"size":500,"mtime":499162500000,"results":"10","hashOfConfig":"11"},{"size":464,"mtime":1611233878131,"results":"12","hashOfConfig":"11"},{"size":362,"mtime":499162500000,"results":"13","hashOfConfig":"11"},{"size":448,"mtime":1611226232827,"results":"14","hashOfConfig":"11"},{"size":320,"mtime":1611245544296,"results":"15","hashOfConfig":"11"},{"size":829,"mtime":1611248339401,"results":"16","hashOfConfig":"11"},{"size":1417,"mtime":1611250454657,"results":"17","hashOfConfig":"11"},{"size":2881,"mtime":1611247224744,"results":"18","hashOfConfig":"11"},{"size":812,"mtime":1611247273588,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},"13beoam",{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\index.js",[],["39","40"],"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\App.js",[],"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\reportWebVitals.js",[],"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\components\\InitialPage\\InitialPage.js",[],"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\components\\Playground\\Playground.js",[],"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\components\\Drums\\Drums.js",[],"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\components\\Mixer\\Mixer.js",[],"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\stores\\audioManager.js",[],"C:\\Users\\Eran\\Desktop\\projects\\virtual-drum-machine\\src\\sounds.js",[],{"ruleId":"41","replacedBy":"42"},{"ruleId":"43","replacedBy":"44"},"no-native-reassign",["45"],"no-negated-in-lhs",["46"],"no-global-assign","no-unsafe-negation"]
