+++ src/components/Drums/Drums.css
@@ -1,0 +1,32 @@
+.drums {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-pack: justify;
+      -ms-flex-pack: justify;
+          justify-content: space-between;
+  -webkit-box-align: baseline;
+      -ms-flex-align: baseline;
+          align-items: baseline;
+  width: 100%;
+  height: 70%;
+  background-image: url("../../assets//images/drum_bg.jpg");
+  background-repeat: no-repeat;
+  background-size: contain;
+  background-position: center;
+}
+
+.drums__title {
+  font-size: 2.5em;
+  font-weight: 400;
+}
+
+.drums__kit-select {
+  width: 200px;
+  height: 30px;
+  padding: 5px;
+  color: #fff;
+  background-color: #470064;
+  border: none;
+}
+/*# sourceMappingURL=Drums.css.map */
