+++ src/components/Drums/Drums.scss
@@ -1,0 +1,26 @@
+.drums {
+  display: flex;
+  justify-content: space-between;
+  align-items: baseline;
+  width: 100%;
+  height: 70%;
+  background-image: url("../../assets//images/drum_bg.jpg");
+  background-repeat: no-repeat;
+  background-size: contain;
+  background-position: center;
+
+  &__title {
+    font-size: 2.5em;
+    font-weight: 400;
+  }
+
+  &__kit-select {
+    width: 200px;
+    height: 30px;
+    // margin: 20px 0;
+    padding: 5px;
+    color: #fff;
+    background-color: #470064;
+    border: none;
+  }
+}
