+++ src/components/Mixer/Mixer.css
@@ -1,0 +1,65 @@
+.mixer {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-pack: start;
+      -ms-flex-pack: start;
+          justify-content: flex-start;
+  width: 100%;
+  height: 30%;
+  padding: 20px 20px;
+  border: 3px solid #eee;
+  border-radius: 12px;
+}
+
+.mixer__title {
+  font-size: 2.5em;
+  font-weight: 400;
+}
+
+.mixer__channels {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-pack: center;
+      -ms-flex-pack: center;
+          justify-content: center;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  width: 100%;
+  height: 100%;
+  margin: 0 auto;
+  overflow-x: auto;
+}
+
+.channel {
+  margin: 10px 20px;
+}
+
+.channel--master {
+  margin-left: 100px;
+}
+
+.channel__fader-wrapper {
+  width: 16px;
+  height: 100px;
+  margin: 5px 20px;
+}
+
+.channel__fader {
+  -webkit-transform: rotate(-90deg);
+          transform: rotate(-90deg);
+  -webkit-transform-origin: 50px 50px;
+          transform-origin: 50px 50px;
+}
+
+.channel__volume-wrapper {
+  width: 100%;
+  text-align: center;
+}
+
+.channel__volume {
+  font-size: 2em;
+}
+/*# sourceMappingURL=Mixer.css.map */
