+++ src/components/Mixer/Mixer.js
@@ -2,6 +2,8 @@ import React, { useContext } from "react";
 import { observer } from "mobx-react";
 import AudioContext from "../../stores/audioManager";
 
+import "./Mixer.css";
+
 function Mixer() {
   const audioManager = useContext(AudioContext);
 
@@ -11,23 +13,32 @@ function Mixer() {
   return (
     <div className='mixer'>
       <h2 className='mixer__title'>mixer</h2>
-      {audioManager.sampleKeys.map(key => (
-        <React.Fragment key={key}>
-          <input
-            // defaultValue={audioManager.initialVolume}
-            className='mixer__fader'
-            type='range'
-            min='0'
-            max='100'
-            step='0.1'
-            name={key}
-            key={key}
-            value={audioManager.getChannelVolume(key)}
-            onChange={e => handleSetVolume(key, e)}
-          />
-          <h3>{audioManager.getChannelVolume(key)}</h3>
-        </React.Fragment>
-      ))}
+      <div className='mixer__channels'>
+        {audioManager.sampleKeys.map(key => (
+          <div className='channel' key={key}>
+            <div className='channel__fader-wrapper'>
+              <div className='channel__fader-wrapper__fader'>
+                <input
+                  className='channel__fader'
+                  type='range'
+                  min='0'
+                  max='100'
+                  step='0.1'
+                  name={key}
+                  key={key}
+                  value={audioManager.getChannelVolume(key)}
+                  onChange={e => handleSetVolume(key, e)}
+                />
+              </div>
+            </div>
+            <div className='channel__volume-wrapper'>
+              <label className='channel__volume'>
+                {audioManager.getChannelVolume(key)}
+              </label>
+            </div>
+          </div>
+        ))}
+      </div>
     </div>
   );
 }
