+++ src/components/Playground/Playground.css
@@ -1,0 +1,12 @@
+.playground {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  width: 95%;
+  height: 95%;
+}
+/*# sourceMappingURL=Playground.css.map */
