+++ src/components/Playground/Playground.scss
@@ -1,0 +1,6 @@
+.playground {
+  display: flex;
+  flex-direction: column;
+  width: 95%;
+  height: 95%;
+}
